requirements.txt
openai-whisper
torch
numpy
pandas
spacy
dateparser
python-dateutil
streamlit